# app/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Yüz Tanıma ile Yoklama Sistemi API",
    "description": "Yüz tanıma teknolojisi kullanarak yoklama almayı ve duygu analizi yapmayı amaçlayan bir sistemin API'si",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
    }
  },
...
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": [
          "🔐 Kimlik Doğrulama"
        ],
        "summary": "Yeni kullanıcı kaydı",
        "description": "Sisteme yeni bir kullanıcı kaydeder",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Kullanıcı bilgileri",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "email",
                "password",
                "first_name",
                "last_name",
                "role"
              ],
              "properties": {
                "email": {
                  "type": "string",
                  "example": "ornek@example.com"
                },
                "password": {
                  "type": "string",
                  "example": "guclu_sifre123"
                },
                "first_name": {
                  "type": "string",
                  "example": "Ahmet"
                },
                "last_name": {
                  "type": "string",
                  "example": "Yılmaz"
                },
                "role": {
                  "type": "string",
                  "enum": [
                    "admin",
                    "teacher",
                    "student"
                  ],
                  "example": "teacher"
                },
                "department": {
                  "type": "string",
                  "example": "Bilgisayar Mühendisliği"
                },
                "title": {
                  "type": "string",
                  "example": "Dr. Öğr. Üyesi"
                },
                "student_number": {
                  "type": "string",
                  "example": "20240001"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Kullanıcı başarıyla kaydedildi",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Kullanıcı başarıyla kaydedildi."
                },
                "user": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Email zaten kayıtlı",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Bu e-posta adresi zaten kullanılıyor."
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "🔐 Kimlik Doğrulama"
        ],
        "summary": "Kullanıcı girişi",
        "description": "Kullanıcı kimlik bilgilerini doğrular ve token döndürür",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Giriş bilgileri",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "email",
                "password"
              ],
              "properties": {
                "email": {
                  "type": "string",
                  "example": "admin@example.com"
                },
                "password": {
                  "type": "string",
                  "example": "admin123"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Başarılı giriş",
            "schema": {
              "type": "object",
              "properties": {
                "access_token": {
                  "type": "string",
                  "example": "eyJhbGciOiJIUzI1NiIsInR5..."
                },
                "refresh_token": {
                  "type": "string",
                  "example": "eyJhbGciOiJIUzI1NiIsInR5..."
                },
                "token_type": {
                  "type": "string",
                  "example": "bearer"
                },
                "user": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Geçersiz kimlik bilgileri",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Geçersiz e-posta adresi veya şifre."
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "🔐 Kimlik Doğrulama"
        ],
        "summary": "Access token yenileme",
        "description": "Refresh token kullanarak yeni bir access token alır",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Token başarıyla yenilendi",
            "schema": {
              "type": "object",
              "properties": {
                "access_token": {
                  "type": "string",
                  "example": "eyJhbGciOiJIUzI1NiIsInR5..."
                }
              }
            }
          },
          "401": {
            "description": "Geçersiz veya süresi dolmuş refresh token",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Geçersiz veya süresi dolmuş token."
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "🔐 Kimlik Doğrulama"
        ],
        "summary": "Mevcut kullanıcı bilgilerini getir",
        "description": "Token ile kimliği doğrulanmış kullanıcının bilgilerini getirir",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Kullanıcı bilgileri başarıyla getirildi",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer",
                  "example": 1
                },
                "email": {
                  "type": "string",
                  "example": "admin@example.com"
                },
                "first_name": {
                  "type": "string",
                  "example": "Admin"
                },
                "last_name": {
                  "type": "string",
                  "example": "User"
                },
                "role": {
                  "type": "string",
                  "example": "admin"
                }
              }
            }
          },
          "404": {
            "description": "Kullanıcı bulunamadı",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Kullanıcı bulunamadı."
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "🔐 Kimlik Doğrulama"
        ],
        "summary": "Mevcut kullanıcı bilgilerini güncelle",
        "description": "Token ile kimliği doğrulanmış kullanıcının bilgilerini günceller",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Güncellenecek kullanıcı bilgileri",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "type": "string",
                  "example": "yeni_sifre123"
                },
                "first_name": {
                  "type": "string",
                  "example": "Yeni Ad"
                },
                "last_name": {
                  "type": "string",
                  "example": "Yeni Soyad"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Kullanıcı bilgileri başarıyla güncellendi",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer",
                  "example": 1
                },
                "email": {
                  "type": "string",
                  "example": "admin@example.com"
                },
                "first_name": {
                  "type": "string",
                  "example": "Yeni Ad"
                },
                "last_name": {
                  "type": "string",
                  "example": "Yeni Soyad"
                },
                "role": {
                  "type": "string",
                  "example": "admin"
                }
              }
            }
          },
          "404": {
            "description": "Kullanıcı bulunamadı",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Kullanıcı bulunamadı."
                }
              }
            }
          },
          "500": {
            "description": "Kullanıcı güncellenemedi",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Kullanıcı güncellenemedi."
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "tags": [
          "🔐 Kimlik Doğrulama"
        ],
        "summary": "Şifre sıfırlama isteği gönder",
        "description": "Kullanıcının şifresini sıfırlamak için bir istek gönderir",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Şifre sıfırlama isteği",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "email"
              ],
              "properties": {
                "email": {
                  "type": "string",
                  "example": "ornek@example.com"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Şifre sıfırlama bağlantısı gönderildi",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Şifre sıfırlama bağlantısı e-posta adresinize gönderildi."
                },
                "token": {
                  "type": "string",
                  "example": "eyJhbGciOiJIUzI1NiIsInR5..."
                }
              }
            }
          },
          "400": {
            "description": "Geçersiz istek",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "E-posta adresi gerekli."
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/reset-password/{token}": {
      "put": {
        "tags": [
          "🔐 Kimlik Doğrulama"
        ],
        "summary": "Şifreyi sıfırla",
        "description": "Verilen token ile kullanıcının şifresini sıfırlar",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "description": "Şifre sıfırlama token'ı",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Yeni şifre",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "password"
              ],
              "properties": {
                "password": {
                  "type": "string",
                  "example": "yeni_sifre123"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Şifre başarıyla sıfırlandı",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Şifreniz başarıyla sıfırlandı."
                }
              }
            }
          },
          "400": {
            "description": "Geçersiz token",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Geçersiz veya süresi dolmuş token."
                }
              }
            }
          },
          "404": {
            "description": "Kullanıcı bulunamadı",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Kullanıcı bulunamadı."
                }
              }
            }
          }
        }
      }
    },
...
  }
} 